Edited screen.h and screen.cpp
Screen is now a struct rather than a class. It no longer deals with
reading from a file (the Editor class will do that); it's task is to
simply display text to the screen, as well as our changes to that text.
As a result, the FILE* data member that Screen previously had is gone.

The display() method now takes a std::vector<std::string> parameter that
represents the contents of the file being edited. Previously it took no
parameters.