Edited editor.h and created editor.cpp
Editor's constructor is now noexcept, and has a new data member
representing the contents of the file being edited. It also has a Screen
member that handles printing text to the screen.

At the time of this commit, the Editor can display the contents of files
to the screen, although its display functionality is very basic.