Added basic cursor movement
Currently the editor will move the cursor in accordance with the vim
bindings in normal mode ('l' will move the cursor right, for example).
At the moment I have no bounds checking on cursor movement, so that will
have to be added later.

The Screen struct now has a method to move the cursor on the screen.

The Editor class now has a method to process kepresses. I haven't
implemented modes at the moment, but this method will have to be changed
to accommodate  vim's modes at some point.

main.cpp now opens "editor.cpp" instead of "makefile", so that way I am
already prepared to test out scrolling (a future feature)

The Cursor struct is now in its own file.