Changed Screen to prepare to implement scrolling
Screen::display() now takes a std::vector<std::string>::iterator rather
than a const std::vector<std::string>&. I did this because I plan to
reuse the display method to implement scrolling, because with an
iterator as the paremeter, I can choose which parts of
Editor::file_contents to display.

Screen now has "rows" and "cols" as data members; they represent the
size of the terminal window. I plan to use these to determine when I
should scroll the screen down.

Screen also now has a new method "show_first_display" that's supposed to
be used when initially displaying a file (in other words, it's supposed
to be called when the editor opens).