Implemented bounded up-down scrolling
Vimperor no longer shows "(null)" when scrolling up and down. It still
has a display bug where it doesn't display the first line of the file; I
plan on tackling this bug in the near future. But as of this commit
Vimperor does a good job of emulating Vim's up-down scrolling (at least
for the Vim on my Linux virtual machine).

As of this commit Editor uses a data member called top_of_screen_index
to determine what the topmost text of the screen should be. It uses this
index for scrolling up and down. For example, when the cursor is at the
bottom of the screen, and the user moves the cursor down,
top_of_screen_index is incremented by 1, meaning that the previous top
line of the screen is now offscreen, and the second-topmost line is now
the topmost line.

I also changed the else branches in the move_cursor_up and
move_cursor_down methods to be else if branches instead, to make the
code less nested.