Fixed Screen::display to work on smaller files
Previously Screen::display would cause Vimperor to crash when opening
smaller files because Screen::display did not check to see if it had
reached the end of the file yet. The "iterator" variable in its for-loop
would keep getting incremented, and would access memory it was not
supposed to access.

Now Screen::display takes 2 iterators so that it can know when it has
displayed the entire file.

I changed the Editor object in main.cpp to open the makefile for this
project instead of editor.cpp to test the new Screen::display (and to
prepare to test additional features, like printing ~'s in the way vim
does)