Made cursor movement and displaying more efficient
Previously, Screen::move_cursor() was always called regardless of
whether the cursor's X and Y positions were changed. This led to a
problem where the cursor could be moved, but the screen wouldn't show
that the cursor was moved in some cases. Moreover, this approach is
inefficient.

This commit has fixed this problem by only calling Screen::move_cursor
when the cursor's X or Y positions have changed.

I also only call Screen::display() only when scrolling or editing is
done for efficiency's sake.