Fixed an out of bounds error in move_cursor_right()
The bug was that if the cursor was on a blank line, it could be moved
right so that an out of bounds error would occur on the check for a tab
character. This bug is now fixed, although now Vimperor doesn't
perfectly mimic Vim's cursor movement (because in Vim, you can only move
the cursor all the way to the last character in a line, but in Vimperor,
you can move one past the last character in a line).