Implemented line deletion
Vimperor can now delete lines as well as characters. However, there is a
bug where deleting on a line after the end of the file will crash the
program.

Also, Vimperor does not totally replicate vim's deletion behavior.
Deleting at the beginning of a line when that line is not empty will
delete characters in the line; in vim this would delete the line and
move the remaining characters to the line above.