Improved cursor behavior when user is typing
The cursor now follows the text being typed in a manner similar to vim.
However, if the current line wraps to another line, the cursor won't be
in the correct position.

The reason why the cursor wasn't following the text before was because
Screen::display() always moves the cursor to the bottom of the screen
because of its calls to printw(). Vimperor needs some refactoring,
because as it stands , whenever we make a call to Screen::display(),
we must also make a call to Screen::move_cursor() AFTER the call to
Screen::display(). Otherwise, the cursor won't be in the correct
position.

The cursor sometimes shows itself at the bottom of the screen as of this
commit due to the repeated calls to Screen::display(). Hopefully, I'll
be able to find a solution to this efficiency problem in the near
future.