Saving files no longer causes the cursor to move
Previously the cursor would move to the inside of the file info bar;
this commit fixes that problem. I had to add a Cursor parameter to
Screen::display() in order to fix this bug, but this meant that
Screen::show_first_display() is unnecessary now (Screen::display() will
work correctly at the start of the program). So
Screen::show_first_display() has been deleted.