Upwards cursor movement is now more like vim's
In vim, whenever the cursor is moved upwards, the cursor might also move
horizontally if the line where the cursor's going is smaller than the
line the cursor was just on. Vimperor now has this functionality for
upwards cursor movement.

This functionality is not perfect, however; tab characters cause Vimperor
some problems.